添加文件asyncStorage，包含对electron-json-storage异步方法的封装，同syncStorage区分
